Refactored Xsts transformation code into three parts
