+++ subprojects/semantifyr/src/main/kotlin/utils/InstanceExtensions.kt
@@ -8,7 +8,7 @@ package hu.bme.mit.semantifyr.oxsts.semantifyr.utils
 
 import hu.bme.mit.semantifyr.oxsts.model.oxsts.Instance
 import hu.bme.mit.semantifyr.oxsts.model.oxsts.impl.InstanceImpl
-import hu.bme.mit.semantifyr.oxsts.semantifyr.transformation.OperationInliner
+import hu.bme.mit.semantifyr.oxsts.semantifyr.transformation.rewrite.OperationInliner
 import hu.bme.mit.semantifyr.oxsts.semantifyr.transformation.evaluation.ContextualExpressionEvaluator
 import hu.bme.mit.semantifyr.oxsts.semantifyr.transformation.evaluation.FeatureEvaluator
 import hu.bme.mit.semantifyr.oxsts.semantifyr.transformation.instantiation.InstancePlacer
